verifier/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
httpx>=0.27.0
google-generativeai>=0.5.0
solana>=0.34.0
solders>=0.21.0
python-dotenv>=1.0.0
